Add custom scalar BigInt to handle large numbers and Infinity

+++ packages/api/schema/types/product/product.js
@@ -123,7 +123,7 @@ export default [
 
     type PriceLevel {
       minQuantity: Int
-      maxQuantity: Int
+      maxQuantity: BigInt
       price: Price
     }
